Rearrange files a bit
