+++ client/web/pinochle/app/game/trick.ts
@@ -1,4 +1,4 @@
-import type { Card } from '../deck';
+import type { Card } from './card';
 
 const POINTS = ['king', 'ace', 10];
