+++ client/web/pinochle/app/game/utils/move-validation.ts
@@ -1,0 +1,6 @@
+import type { Card, Suit } from '../card';
+import type { Trick } from '../trick';
+
+export function isValidMove(trick: Trick, card: Card, trump: Suit) {}
+
+export function availableMoves(trick: Trick, hand: Card[], trump: Suit) {}
