--- client/web/pinochle/app/utils/game/move-validation.ts
@@ -1,6 +1,0 @@
-import type { Card, Suit } from '../deck';
-import type { Trick } from 'pinochle/game/trick';
-
-export function isValidMove(trick: Trick, card: Card, trump: Suit) {
-
-}
