+++ client/web/pinochle/tests/unit/game/utils/availableMoves-test.ts
@@ -1,0 +1,5 @@
+import { module, test } from 'qunit';
+
+import { availableMoves } from 'pinochle/game/utils/move-validation';
+
+module('Unit | Game | Utils | availableMoves', function () {});
