+++ client/web/pinochle/tests/unit/game/utils/isValidMove-test.ts
@@ -1,0 +1,9 @@
+import { module, test } from 'qunit';
+
+import { isValidMove } from 'pinochle/game/utils/move-validation';
+
+module('Unit | Game | Utils | isValidMove', function () {
+  test('must match suit', function () {
+    let trick = new Trick(3);
+  });
+});
